# crates/synthia-server/docs/api-reference/openapi.yaml
openapi: 3.0.3
info:
  title: Synthia Server API
  description: |
    Synthia Server 是 Synthia Agent 的 HTTP API 服务器，为前端应用、编辑器插件和 TUI 客户端提供 REST API 接口。
    
    ## 认证
    
    当认证启用时，所有请求需要在 `Authorization` 头中携带 Bearer Token：
    ```
    Authorization: Bearer <api_key>
    ```
    
    ## 响应格式
    
    所有响应使用 JSON 格式，成功响应包含 `data` 字段，错误响应包含 `error` 字段。
    
    ## 流式响应
    
    `/chat/stream` 端点使用 Server-Sent Events (SSE) 返回流式响应。
    
    ## WebSocket
    
    `/ws/{session_id}` 端点提供 WebSocket 连接用于实时双向通信。
  version: 1.0.0
  contact:
    name: Synthia Team
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发服务器
...
paths:
...
  /ws/{session_id}:
    get:
      tags:
        - Chat
      summary: WebSocket聊天
      description: |
        建立WebSocket连接进行实时聊天。
        
        ## 客户端发送消息格式
        
        发送聊天消息：
        ```json
        {"action": "chat", "content": "你好"}
        ```
        
        取消当前操作：
        ```json
        {"action": "cancel"}
        ```
        
        ## 服务端推送消息格式
        
        消息事件：
        ```json
        {"type": "message", "content": "..."}
        ```
        
        状态事件：
        ```json
        {"type": "status", "status": "Completed"}
        ```
        
        错误事件：
        ```json
        {"type": "error", "error": "..."}
        ```
      operationId: websocket
      parameters:
        - name: session_id
          in: path
          required: true
          description: 会话ID
          schema:
            type: string
      responses:
        '101':
          description: 切换到WebSocket协议
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API Key认证
security:
  - BearerAuth: []